# Enabling 'fixed-length2' schema in 'omni.2.1' schema handler; adding samples and creating parity comparison against original 'fixed-length' and their benchmarks (no degradation); adding 'fixed-length2' to heroku server sample collection.