--- a/cli/cmd/serverCmd.go
+++ b/cli/cmd/serverCmd.go
@@ -154,7 +154,7 @@ func httpPostTransform(w http.ResponseWriter, r *http.Request) {
 
 var (
 	sampleDir                  = "../../extensions/omniv21/samples/"
-	sampleFormats              = []string{"csv", "json", "xml", "fixedlength", "edi"}
+	sampleFormats              = []string{"csv", "json", "xml", "fixedlength", "fixedlength2", "edi"}
 	sampleInputFilenamePattern = regexp.MustCompile("^([0-9]+[_a-zA-Z0-9]+)\\.input\\.[a-z]+$")
 )