--- a/extensions/omniv21/samples/fixedlength2/2_multi_rows.input.txt
+++ b/extensions/omniv21/samples/fixedlength2/2_multi_rows.input.txt
@@ -1,0 +1,22 @@
+H001TW                  0689311345         030                           DEL       HMDEPOT                                                   20190826124704  EST                                                   1230004321         W841206858                                                                                                         James Bond                    20190827                                                  RHONDA                        W841206858                    128 Bird Ave                                                          HAPPYVALLEY                   FL54321      USELGS               030                                                     1                                       105.00   THDNR     1    1003164515          DLB000977979                                      Appt set between 00 00 AND 00 00 on 08 27 19                                                                                                                                                                                                                                                                AG                                                                                      21200    1600    7029083C                                                                                    314 Algebra Blvd                                            MEDIAN              FL31415  US
+H0020689311345         BM  040670500067120
+H0020689311345         CN  100000103732
+H0020689311345         OA  2400 Highway 155 South
+H0020689311345         OC  Locust Grove
+H0020689311345         ON  THE HOME DEPOT 6705
+H0020689311345         OO  W841206858
+H0020689311345         OP  30248
+H0020689311345         OST GA
+H0020689311345         OT  US
+H0020689311345         PO  7029083C
+H001TW                  0689311348         030                           DEL       HMDEPOT                                                   20190826124704  EST                                                   1230001234         W938003272                                                                                                         Jason Bourne                  20190827                                                  RHONDA                        W938003272                    123 S 45ST ST                                                         MAGIC BEACH                   FL12345      USELGS               030                                                     1                                       1.00     THDNR     1    1003621621                                                            Appt set between 00 00 AND 00 00 on 08 27 19                                                                                                                                                                                                                                                                AG                                                                                      21030    1430    6816137C                                                                                    314 Algebra Blvd                                            MEDIAN              FL31415  US
+H0020689311348         BM  040677700277714
+H0020689311348         CN
+H0020689311348         OA  2500 Highway 155 South
+H0020689311348         OC  Locust Grove
+H0020689311348         ON  THE HOME DEPOT 6777
+H0020689311348         OO  W938003272
+H0020689311348         OP  30248
+H0020689311348         OST GA
+H0020689311348         OT  US
+H0020689311348         PO  6816137C