--- a/extensions/omniv21/samples/fixedlength2/fixedlength_test.go
+++ b/extensions/omniv21/samples/fixedlength2/fixedlength_test.go
@@ -1,4 +1,4 @@
-package fixedlength
+package fixedlength2
 
 import (
 	"bytes"
@@ -25,6 +25,7 @@ const (
 	test1_Single_Row = iota
 	test2_Multi_Rows
 	test3_Header_Footer
+	test4_Nested
 )
 
 var tests = []testCase{
@@ -38,10 +39,16 @@ var tests = []testCase{
 		schemaFile: "./2_multi_rows.schema.json",
 		inputFile:  "./2_multi_rows.input.txt",
 	},
+
 	{
 		// test3_Header_Footer
 		schemaFile: "./3_header_footer.schema.json",
 		inputFile:  "./3_header_footer.input.txt",
+	},
+	{
+		// test4_Nested
+		schemaFile: "./4_nested.schema.json",
+		inputFile:  "./4_nested.input.txt",
 	},
 }
 
@@ -97,17 +104,26 @@ func Test3_Header_Footer(t *testing.T) {
 	tests[test3_Header_Footer].doTest(t)
 }
 
-// Benchmark1_Single_Row-8      	   25869	     45576 ns/op	   27721 B/op	     644 allocs/op
+func Test4_Nested(t *testing.T) {
+	tests[test4_Nested].doTest(t)
+}
+
+// Benchmark1_Single_Row-8      	   25898	     45728 ns/op	   28169 B/op	     647 allocs/op
 func Benchmark1_Single_Row(b *testing.B) {
 	tests[test1_Single_Row].doBenchmark(b)
 }
 
-// Benchmark2_Multi_Rows-8      	   18813	     63901 ns/op	   29167 B/op	     635 allocs/op
+// Benchmark2_Multi_Rows-8      	   15338	     78273 ns/op	   30179 B/op	     618 allocs/op
 func Benchmark2_Multi_Rows(b *testing.B) {
 	tests[test2_Multi_Rows].doBenchmark(b)
 }
 
-// Benchmark3_Header_Footer-8   	    5857	    197326 ns/op	   82234 B/op	    2009 allocs/op
+// Benchmark3_Header_Footer-8   	    6390	    178301 ns/op	   76571 B/op	    1501 allocs/op
 func Benchmark3_Header_Footer(b *testing.B) {
 	tests[test3_Header_Footer].doBenchmark(b)
+}
+
+// Benchmark4_Nested-8          	   10000	    107948 ns/op	   78986 B/op	    1535 allocs/op
+func Benchmark4_Nested(b *testing.B) {
+	tests[test4_Nested].doBenchmark(b)
 }