old `fixed-length` benches:
```
// Benchmark1_Single_Row-8      	   25869	     45576 ns/op	   27721 B/op	     644 allocs/op
// Benchmark2_Multi_Rows-8      	   18813	     63901 ns/op	   29167 B/op	     635 allocs/op
// Benchmark3_Header_Footer-8   	    5857	    197326 ns/op	   82234 B/op	    2009 allocs/op
```

new `fixed-length2` benches:
```
// Benchmark1_Single_Row-8      	   25898	     45728 ns/op	   28169 B/op	     647 allocs/op
// Benchmark2_Multi_Rows-8      	   15338	     78273 ns/op	   30179 B/op	     618 allocs/op
// Benchmark3_Header_Footer-8   	    6390	    178301 ns/op	   76571 B/op	    1501 allocs/op
// Benchmark4_Nested-8          	   10000	    107948 ns/op	   78986 B/op	    1535 allocs/op
```

As you can see there is no performance degradation while `fixed-length2` added significant feature upgrade to the schema. `fixed-length2` features are strictly a super set of the old `fixed-length`. As such we'll soon be deprecating `fixed-length` - that is, while keeping it functional, tested, and maintained, we will start to phase it out from main documentation and will add a migration strategy from it the new `2` schema - while `2` vs old schema-wise isn't strictly back-compatible, the migration/change needed is incredibly trivial.